requirements.txt
segment-anything @ git+https://github.com/facebookresearch/segment-anything
flask
pillow
numpy
requests
pywin32